booked service added

+++ src/Pages/Manage/Manage.jsx
@@ -4,6 +4,7 @@ import { FaRegEdit } from "react-icons/fa";
 import { FiXCircle } from "react-icons/fi";
 import toast from "react-hot-toast";
 import axios from "axios";
+import Swal from "sweetalert2";
 
 const Manage = () => {
   const [bookData, setBookData] = useState([]);
@@ -15,6 +16,7 @@ const Manage = () => {
 
 
   const getData = () => {
+    // to do : which data she / he added 
     fetch(`http://localhost:5000/bookedServices/${user?.email}`)
       .then((res) => res.json())
       .then((data) => {
@@ -27,13 +29,54 @@ const Manage = () => {
     return <p> loading...... </p>;
   }
 
-  const handleDelete = async (id) => {
+  const handleDelete = (id) => {
     console.log(id);
     try{
-      const data = await axios.delete(`http://localhost:5000/bookedServices/${id}`)
+
+
+
+
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+
+
+          const data = axios.delete(`http://localhost:5000/bookedServices/${id}`)
       console.log(data);
       toast.success('delete successful');
       getData();
+
+
+
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success"
+          });
+        }
+      });
+
+
+
+
+      // const data = await axios.delete(`http://localhost:5000/bookedServices/${id}`)
+      // console.log(data);
+      // toast.success('delete successful');
+      // getData();
+
+
+
+
+
+
+      
     }
     catch{err=>{
       toast.error(err.message)
